index.html: Highlight the selected link
Instead of using target attribute, we add an event listener on_click() to
highlight the selected link and load corresponding page in iframe. To
tackle with the case of page reload, we save previous selected link
in the cookie, and read it back during reload.

It works with Chrome, but somehow doesn't with Firefox.