Add queue web page
The page is used for party staff to understand current status of prize queue.
In the end of each row there is a button used for confirm that the prize has
been received. The web page refreshes content every 5 secs, the prize having
been received will be removed from the queue.