donator.html: Add a field to show amount to be doanted
According to different price of each prize, the amount the donator needs to
donate is different. We add a new field to show amount to be donated, which
has better overview.

The view.py is also updated to return a new field 'donate_amount' in JSON
data. Note the amount is hard-coded here, and will be changed later.