queue.js: Don't just overwrite existing model
Since before committing to database, a dialog appears to prompt user to
click comfirm button, there is a duration that the system is waiting for
user's response. But the model updating still continues every 5 secs.
Directly overwritting the model may cause the item referenced by the
function confirm_received() no longer appearing in web page, and the
status message won't be shown even if committing is successful.

So instead of just overwritting, we should compare the queried data &
current model, and update the model accordingly.