employee.js: Add a small delay before querying
In fuzzy jobid query mode, the callback executes query immediately when the
user is typing, which is a meaningless operation, because the returned data
is not what the user expect, and it's easy to cause race condition in model
update. So we add a small delay to wait the user completing his typing, and
compare the argument with current inputted text to decide if query needs to
be executed. After experiment, a 400ms delay is appropriate.