queue.js: Re-loop the array after removing elements
After removing an element in the array, the indices of elements after the
one being removed will change, so only keeping indices information is not
a good idea when there are multiple elements to be removed. We should loop
into the array to find correct index every time before removing an element
from the array.